address review (add a lot of code comments) & some nits
